Working CSS and HTML files
The JS files are completely broken and not working
I have most of the basic js in index.js. Everything
else is everywhere. I still need to plan how to
allocate my code to the correct place and do the
css for dark mode.

This is my first time on scss. I think I need to dl
npm for it.